requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
arrow==1.2.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.1
blessed==1.19.1
cachetools==5.3.0
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
croniter==1.3.8
deepdiff==6.2.2
dnspython==2.2.1
email-validator==1.3.0
exceptiongroup==1.1.0
fastapi==0.88.0
filelock==3.9.0
frozenlist==1.3.3
fsspec==2022.11.0
func-timeout==4.3.5
google-auth==2.16.2
google-auth-oauthlib==0.4.6
gpustat==1.0.0
grpcio==1.51.3
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.1
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==6.1.0
inquirer==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
lightning==1.8.6
lightning-cloud==0.5.15
lightning-utilities==0.5.0
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.4
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py==11.495.46
oauthlib==3.2.2
ordered-set==4.1.0
orjson==3.8.3
outcome==1.2.0
packaging==22.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.20.1
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.17
pydantic==1.10.4
Pygments==2.13.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-editor==1.0.4
python-multipart==0.0.5
pytorch-lightning==1.8.6
pytz==2022.7.1
PyYAML==6.0
readchar==4.0.3
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.0.0
rsa==4.9
schedule==1.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
starlette==0.22.0
starsessions==1.3.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tokenizers==0.13.2
torch==1.13.1
torch-tb-profiler==0.4.1
torchmetrics==0.11.0
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.8.0
transformers==4.25.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
ujson==5.6.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
wcwidth==0.2.5
Werkzeug==2.2.3
wsproto==1.2.0
yarl==1.8.2
zipp==3.15.0